+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Clean `--bin` directory before installing a new version of the dependency-check-cli.
 - Added parameter `--keep-old-versions` to prevent removal of other installations in `--bin` directory.
+- The dependency `@colors/colors` has been replaced with `ansis`.
 
 ## Version 0.7.0