+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unrelease Changes
+## Unreleased Changes
 
 - Clean `--bin` directory before installing a new version of the dependency-check-cli.
 - Added parameter `--keep-old-versions` to prevent removal of other installations in `--bin` directory.