+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 # Changelog
 
+## Unreleased Changes
+
+- Clean `--bin` directory before installing a new version of the dependency-check-cli.
+- Added parameter `--keep-old-versions` to prevent removal of other installations in `--bin` directory.
+- The dependency `@colors/colors` has been replaced with `ansis`.
+- The dependency `commander` has been upgraded to version 14.
+
 ## Version 0.7.1
 
 - Use `rollup` and `terser` to create a single file compressed release version.
 - Removed `CHANGELOG.md` from release package.
-
+- 
 ## Version 0.7.0
 
 - The output of the `--version` parameter was fixed. It now displays the version of owasp-dependency-check instead of the version of the analyzed project.